// src/infrastructure/graphql/generated-types.ts
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type User = {
  __typename?: 'User';
  username: Scalars['String'];
};

export type UsernameAlreadyTakenError = Error & {
  __typename?: 'UsernameAlreadyTakenError';
  message: Scalars['String'];
};

export type UserNotFoundError = Error & {
  __typename?: 'UserNotFoundError';
  message: Scalars['String'];
};

export type RefreshAccessTokenSuccessPayload = {
  __typename?: 'RefreshAccessTokenSuccessPayload';
  accessToken: Scalars['String'];
};

export type RefreshAccessTokenPayload = RefreshAccessTokenSuccessPayload | UserNotFoundError;

export type SignInInput = {
  password: Scalars['String'];
  username: Scalars['String'];
};

export type SignInSuccessPayload = {
  __typename?: 'SignInSuccessPayload';
  accessToken: Scalars['String'];
};

export type SignInPayload = SignInSuccessPayload | UserNotFoundError;

export type SignOutSuccessPayload = {
  __typename?: 'SignOutSuccessPayload';
  _empty?: Maybe<Scalars['String']>;
};

export type SignOutPayload = SignOutSuccessPayload | UserNotFoundError;

export type SignUpInput = {
  username: Scalars['String'];
  password: Scalars['String'];
};

export type SignUpSuccessPayload = {
  __typename?: 'SignUpSuccessPayload';
  user?: Maybe<User>;
};

export type SignUpPayload = SignUpSuccessPayload | UsernameAlreadyTakenError;

export type Mutation = {
  __typename?: 'Mutation';
  refreshAccessToken: RefreshAccessTokenPayload;
  signIn: SignInPayload;
  signOut: SignOutPayload;
  signUp: SignUpPayload;
};


export type MutationSignInArgs = {
  input: SignInInput;
};


export type MutationSignUpArgs = {
  input: SignUpInput;
};

export type Query = {
  __typename?: 'Query';
  me?: Maybe<User>;
};

export type Error = {
  message: Scalars['String'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;

export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = {
  User: ResolverTypeWrapper<User>;
  String: ResolverTypeWrapper<Scalars['String']>;
  UsernameAlreadyTakenError: ResolverTypeWrapper<UsernameAlreadyTakenError>;
  UserNotFoundError: ResolverTypeWrapper<UserNotFoundError>;
  RefreshAccessTokenSuccessPayload: ResolverTypeWrapper<RefreshAccessTokenSuccessPayload>;
  RefreshAccessTokenPayload: ResolversTypes['RefreshAccessTokenSuccessPayload'] | ResolversTypes['UserNotFoundError'];
  SignInInput: SignInInput;
  SignInSuccessPayload: ResolverTypeWrapper<SignInSuccessPayload>;
  SignInPayload: ResolversTypes['SignInSuccessPayload'] | ResolversTypes['UserNotFoundError'];
  SignOutSuccessPayload: ResolverTypeWrapper<SignOutSuccessPayload>;
  SignOutPayload: ResolversTypes['SignOutSuccessPayload'] | ResolversTypes['UserNotFoundError'];
  SignUpInput: SignUpInput;
  SignUpSuccessPayload: ResolverTypeWrapper<SignUpSuccessPayload>;
  SignUpPayload: ResolversTypes['SignUpSuccessPayload'] | ResolversTypes['UsernameAlreadyTakenError'];
  Mutation: ResolverTypeWrapper<{}>;
  Query: ResolverTypeWrapper<{}>;
  Error: ResolversTypes['UsernameAlreadyTakenError'] | ResolversTypes['UserNotFoundError'];
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  User: User;
  String: Scalars['String'];
  UsernameAlreadyTakenError: UsernameAlreadyTakenError;
  UserNotFoundError: UserNotFoundError;
  RefreshAccessTokenSuccessPayload: RefreshAccessTokenSuccessPayload;
  RefreshAccessTokenPayload: ResolversParentTypes['RefreshAccessTokenSuccessPayload'] | ResolversParentTypes['UserNotFoundError'];
  SignInInput: SignInInput;
  SignInSuccessPayload: SignInSuccessPayload;
  SignInPayload: ResolversParentTypes['SignInSuccessPayload'] | ResolversParentTypes['UserNotFoundError'];
  SignOutSuccessPayload: SignOutSuccessPayload;
  SignOutPayload: ResolversParentTypes['SignOutSuccessPayload'] | ResolversParentTypes['UserNotFoundError'];
  SignUpInput: SignUpInput;
  SignUpSuccessPayload: SignUpSuccessPayload;
  SignUpPayload: ResolversParentTypes['SignUpSuccessPayload'] | ResolversParentTypes['UsernameAlreadyTakenError'];
  Mutation: {};
  Query: {};
  Error: ResolversParentTypes['UsernameAlreadyTakenError'] | ResolversParentTypes['UserNotFoundError'];
  Boolean: Scalars['Boolean'];
};
// ...
export type MutationResolvers<ContextType = GraphQLContext, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  refreshAccessToken?: Resolver<ResolversTypes['RefreshAccessTokenPayload'], ParentType, ContextType>;
  signIn?: Resolver<ResolversTypes['SignInPayload'], ParentType, ContextType, RequireFields<MutationSignInArgs, 'input'>>;
  signOut?: Resolver<ResolversTypes['SignOutPayload'], ParentType, ContextType>;
  signUp?: Resolver<ResolversTypes['SignUpPayload'], ParentType, ContextType, RequireFields<MutationSignUpArgs, 'input'>>;
};